rm footer.v1
