+++ src/components/roar/landing-page/index.tsx
@@ -17,7 +17,7 @@ const RoarPage = ({ location }: PageProps): JSX.Element => {
   useExplicitHeightOnIPhone(internalSectionRefs['hero'], internalSectionRefs['How it works']) // tslint:disable-line:no-expression-statement
 
   return (
-    <Layout additionalClassNames="roar" footerKind="v2" header={header}>
+    <Layout additionalClassNames="roar" header={header}>
       <RoarSEO />
       <Hero ref={internalSectionRefs['hero']} dotsRef={dotsRef} />
       <HowItWorks ref={internalSectionRefs['How it works']} />
