+++ src/components/shared/footer/built-with.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import MHIDotsLogo from '../../mhi-dots-logo'
+import MHIDotsLogo from '../mhi-dots-logo'
 
 
 const BuiltWith = (): JSX.Element => {
