+++ src/components/shared/footer/index.tsx
@@ -1,11 +1,48 @@
 import React from 'react'
-import V1 from './v1'
-import V2 from './v2'
+import BuiltWith from './built-with'
+import { BuyMeACoffeeIcon, EmailIcon, GithubIcon, LinkedInIcon, TwitterIcon } from './social-icons'
+import SocialLinks from './social-links'
 
-export default function Footer({ kind }: { kind: 'v1' | 'v2' }): JSX.Element {
-  if (kind === 'v1') {
-    return <V1 />
-  } else {
-    return <V2 />
-  }
+const socialLinkData: ReadonlyArray<Link> = [
+  {
+    href: 'https://github.com/morehumaninternet',
+    ariaLabel: 'Github',
+    icon: <GithubIcon />,
+  },
+  {
+    href: 'https://twitter.com/morehumaninter1',
+    ariaLabel: 'Twitter',
+    icon: <TwitterIcon />,
+  },
+  {
+    href: 'https://www.linkedin.com/company/more-human-internet/',
+    ariaLabel: 'LinkedIn',
+    icon: <LinkedInIcon />,
+  },
+  {
+    href: 'https://www.buymeacoffee.com/morehumaninter',
+    ariaLabel: 'Buy Me A Coffee',
+    icon: <BuyMeACoffeeIcon />,
+  },
+  {
+    href: 'mailto:info@morehumaninternet.org',
+    ariaLabel: 'Email',
+    icon: <EmailIcon />,
+  },
+]
+
+const Footer = (): JSX.Element => {
+  return (
+    <footer className="v2">
+      <BuiltWith />
+      <div className="footer-banner">
+        <div className="footer-links__container">
+          <SocialLinks links={socialLinkData} />
+        </div>
+        <span className="copyright">© More Human Internet Foundation 2020, All rights reserved</span>
+      </div>
+    </footer>
+  )
 }
+
+export default Footer
