+++ src/styles/components/footer.scss
@@ -1,4 +1,4 @@
-footer.v2 {
+footer {
   font-size: 0.8em;
   padding-top: 2.5rem;
